+++ src/main/java/uk/ac/ebi/atlas/bioentity/properties/BioEntityPropertyService.java
@@ -6,7 +6,7 @@ import uk.ac.ebi.atlas.bioentity.go.GoPoTrader;
 import uk.ac.ebi.atlas.bioentity.interpro.InterProTrader;
 import uk.ac.ebi.atlas.model.OntologyTerm;
 import uk.ac.ebi.atlas.search.SemanticQuery;
-import uk.ac.ebi.atlas.solr.BioentityPropertyName;
+import uk.ac.ebi.atlas.solr.bioentities.BioentityPropertyName;
 import uk.ac.ebi.atlas.species.SpeciesInferrer;
 import uk.ac.ebi.atlas.utils.ReactomeClient;
 
@@ -16,9 +16,9 @@ import java.util.Map;
 import static java.util.function.Function.identity;
 import static java.util.stream.Collectors.toMap;
 import static org.apache.commons.text.WordUtils.capitalize;
-import static uk.ac.ebi.atlas.solr.BioentityPropertyName.GO;
-import static uk.ac.ebi.atlas.solr.BioentityPropertyName.PO;
-import static uk.ac.ebi.atlas.solr.BioentityPropertyName.SYMBOL;
+import static uk.ac.ebi.atlas.solr.bioentities.BioentityPropertyName.GO;
+import static uk.ac.ebi.atlas.solr.bioentities.BioentityPropertyName.PO;
+import static uk.ac.ebi.atlas.solr.bioentities.BioentityPropertyName.SYMBOL;
 
 @Service
 public class BioEntityPropertyService {
