+++ src/main/java/uk/ac/ebi/atlas/solr/bioentities/query/BioentitiesSolrClient.java
@@ -9,7 +9,7 @@ import org.apache.solr.common.SolrDocument;
 import org.apache.solr.common.SolrException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import uk.ac.ebi.atlas.solr.BioentityPropertyName;
+import uk.ac.ebi.atlas.solr.bioentities.BioentityPropertyName;
 
 import javax.inject.Inject;
 import javax.inject.Named;
