+++ src/main/java/uk/ac/ebi/atlas/web/BaselineRequestPreferences.java
@@ -3,7 +3,7 @@ package uk.ac.ebi.atlas.web;
 import uk.ac.ebi.atlas.model.ExpressionUnit;
 import uk.ac.ebi.atlas.search.SemanticQuery;
 import uk.ac.ebi.atlas.search.SemanticQueryTerm;
-import uk.ac.ebi.atlas.solr.BioentityPropertyName;
+import uk.ac.ebi.atlas.solr.bioentities.BioentityPropertyName;
 
 public abstract class BaselineRequestPreferences<U extends ExpressionUnit.Absolute>
         extends ExperimentPageRequestPreferences<U> {
